requirements.txt
discord.py>=2.0.0
python-dotenv
Pillow>=10.0.0
